Change: Entire AterraCore.Boot rework
Twitch Redeems
- Brought to you by ToRa_Pan with "sorry i'm late"
- Brought to you by varis_revolver

+++ src/engine/AterraEngine/Engine.cs
@@ -95,6 +95,6 @@ public class Engine(
     private async Task HandleFatalExceptionGracefully() {
         await _ctsRenderThread.CancelAsync();
 
-        logger.ExitFatal(ExitCodes.GeneralError, "Fatally Crashing gracefully");
+        logger.ExitFatal((int)ExitCodes.GeneralError, "Fatally Crashing gracefully");
     }
 }
